// src/app/shared/layout/models/company-info.model.ts
// tslint:disable: variable-name
export class CompanyInfoModel {
    address: string;
    email: string;
    name: string;
    phone: string;
    url: string;
    api_url: string;
    branch_id: string;
    decimals_for_money: number;
    default_location: number;
    decimals_in_general: number;
    decimals_in_tickets: number;
    decimal_for_characteristics: number;
    metric_system_id: number;
    mstatus: number;
    source_id: string;
    money: number;
    ticket: number;
    general: number;
    metric: number;
    characteristics: number;
    id: any;
    sqs_arn: string;
    sqs_key: string;
    sqs_name: string;
    sqs_secret: string;
    sqs_url: string;
}